feat(tests): ✨ Add `StorageActionTests` for file upload and download functionality; Includes tests for successful file upload and download operations; Removes unnecessary package references
